requirements.txt
openai==0.28
pandas==2.2.3
pyyaml==6.0.2
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20